Create: Semi-final Triage System
Triage is almost completed, we only need to fix where the Summary will be send.

Fix the Images and their Descriptions/names

And after that, we are finished!

I Re-organized the Triage.HTML, now there is a Triage JS, CSS and HTML (Separated files).

Debug Prints is disabled for now.